# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: GBJ SolarDAMS API — Phase 1 (Asset Registry, Maintenance, Customer Portal, Housing Developer)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all sites
      responses:
        "200":
          description: Site list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{siteId}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a site by ID
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update a site
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PV Assets ───────────────────────────────────────────────
...
  /assets/{assetId}/equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment for an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Register equipment on an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{equipmentId}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Work Orders ─────────────────────────────────────────────
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [workorders]
      summary: List all work orders
      responses:
        "200":
          description: Work order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [workorders]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{workOrderId}:
    get:
      operationId: getWorkOrder
      tags: [workorders]
      summary: Get a work order by ID
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          description: Not found
    patch:
      operationId: updateWorkOrder
      tags: [workorders]
      summary: Update a work order
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [workorders]
      summary: Delete a work order
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Housing Developer Projects ───────────────────────────────
...
  /developer-projects/{projectId}/units:
    get:
      operationId: listHousingUnits
      tags: [units]
      summary: List housing units for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HousingUnit"
    post:
      operationId: createHousingUnit
      tags: [units]
      summary: Create a housing unit
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HousingUnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HousingUnit"
  /housing-units/{unitId}:
    get:
      operationId: getHousingUnit
      tags: [units]
      summary: Get a housing unit by ID
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HousingUnit"
        "404":
          description: Not found
    patch:
      operationId: updateHousingUnit
      tags: [units]
      summary: Update a housing unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HousingUnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HousingUnit"
    delete:
      operationId: deleteHousingUnit
      tags: [units]
      summary: Delete a housing unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Site:
      type: object
      required: [id, siteName, siteCode, address, siteType, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        siteName:
          type: string
        siteCode:
          type: string
        address:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        siteType:
          type: string
          enum: [residential, commercial, industrial, utility_scale, government]
        gridRegion:
          type: ["string", "null"]
        utilityProvider:
          type: ["string", "null"]
        commissioningDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, under_construction, decommissioned]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    SiteInput:
      type: object
      required: [siteName, siteCode, address, siteType, status]
      properties:
        customerId:
          type: integer
        siteName:
          type: string
          minLength: 1
        siteCode:
          type: string
          minLength: 1
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        siteType:
          type: string
          enum: [residential, commercial, industrial, utility_scale, government]
        gridRegion:
          type: string
        utilityProvider:
          type: string
        commissioningDate:
          type: string
        status:
          type: string
          enum: [active, inactive, under_construction, decommissioned]
    SiteUpdate:
      type: object
      properties:
        customerId:
          type: integer
        siteName:
          type: string
        siteCode:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        siteType:
          type: string
          enum: [residential, commercial, industrial, utility_scale, government]
        gridRegion:
          type: string
        utilityProvider:
          type: string
        commissioningDate:
          type: string
        status:
          type: string
          enum: [active, inactive, under_construction, decommissioned]

    # ─── PV Asset ─────────────────────────────────────────
...
    Equipment:
      type: object
      required: [id, assetId, equipmentType, manufacturer, model, serialNumber, status, createdAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        equipmentType:
          type: string
          enum: [pv_module, inverter, battery, meter, switchboard, mounting_system, sensor, communications, other]
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        ratedCapacity:
          type: ["string", "null"]
        installationDate:
          type: ["string", "null"]
        warrantyStart:
          type: ["string", "null"]
        warrantyEnd:
          type: ["string", "null"]
        locationReference:
          type: ["string", "null"]
        status:
          type: string
          enum: [operational, faulty, replaced, decommissioned]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    EquipmentInput:
      type: object
      required: [equipmentType, manufacturer, model, serialNumber, status]
      properties:
        equipmentType:
          type: string
          enum: [pv_module, inverter, battery, meter, switchboard, mounting_system, sensor, communications, other]
        manufacturer:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        serialNumber:
          type: string
          minLength: 1
        ratedCapacity:
          type: string
        installationDate:
          type: string
        warrantyStart:
          type: string
        warrantyEnd:
          type: string
        locationReference:
          type: string
        status:
          type: string
          enum: [operational, faulty, replaced, decommissioned]
    EquipmentUpdate:
      type: object
      properties:
        equipmentType:
          type: string
          enum: [pv_module, inverter, battery, meter, switchboard, mounting_system, sensor, communications, other]
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        ratedCapacity:
          type: string
        installationDate:
          type: string
        warrantyStart:
          type: string
        warrantyEnd:
          type: string
        locationReference:
          type: string
        status:
          type: string
          enum: [operational, faulty, replaced, decommissioned]

    # ─── Work Order ───────────────────────────────────────
    WorkOrder:
      type: object
      required: [id, assetId, workOrderNumber, type, priority, description, status, createdAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        workOrderNumber:
          type: string
        type:
          type: string
          enum: [preventive, corrective, emergency, inspection, warranty]
        priority:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        assignedTo:
          type: ["string", "null"]
        scheduledDate:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, open, assigned, scheduled, in_progress, awaiting_parts, awaiting_approval, completed, verified, closed, cancelled]
        labourCost:
          type: ["number", "null"]
        materialCost:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        findings:
          type: ["string", "null"]
        correctiveAction:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    WorkOrderInput:
      type: object
      required: [assetId, type, priority, description, status]
      properties:
        assetId:
          type: integer
        type:
          type: string
          enum: [preventive, corrective, emergency, inspection, warranty]
        priority:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
          minLength: 1
        assignedTo:
          type: string
        scheduledDate:
          type: string
        status:
          type: string
          enum: [draft, open, assigned, scheduled, in_progress, awaiting_parts, awaiting_approval, completed, verified, closed, cancelled]
    WorkOrderUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [preventive, corrective, emergency, inspection, warranty]
        priority:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        assignedTo:
          type: string
        scheduledDate:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        status:
          type: string
          enum: [draft, open, assigned, scheduled, in_progress, awaiting_parts, awaiting_approval, completed, verified, closed, cancelled]
        labourCost:
          type: number
        materialCost:
          type: number
        findings:
          type: string
        correctiveAction:
          type: string

    # ─── Developer Project ────────────────────────────────
...
    HousingUnit:
      type: object
      required: [id, projectId, unitNumber, unitStatus, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        unitNumber:
          type: string
        block:
          type: ["string", "null"]
        phase:
          type: ["string", "null"]
        purchaserName:
          type: ["string", "null"]
        purchaserContact:
          type: ["string", "null"]
        pvCapacityKwp:
          type: ["number", "null"]
        inverterModel:
          type: ["string", "null"]
        inverterSerial:
          type: ["string", "null"]
        moduleModel:
          type: ["string", "null"]
        moduleSerial:
          type: ["string", "null"]
        utilityAccountNumber:
          type: ["string", "null"]
        meterNumber:
          type: ["string", "null"]
        nemApplicationDate:
          type: ["string", "null"]
        nemApprovalDate:
          type: ["string", "null"]
        commissioningDate:
          type: ["string", "null"]
        handoverDate:
          type: ["string", "null"]
        warrantyStart:
          type: ["string", "null"]
        warrantyEnd:
          type: ["string", "null"]
        unitStatus:
          type: string
          enum: [not_started, design_confirmed, equipment_allocated, installation_in_progress, installation_completed, testing_completed, utility_submission_pending, utility_submission_submitted, utility_approval_received, meter_installed, commissioned, handover_pending, handed_over, under_warranty, closed]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    HousingUnitInput:
      type: object
      required: [unitNumber, unitStatus]
      properties:
        unitNumber:
          type: string
          minLength: 1
        block:
          type: string
        phase:
          type: string
        purchaserName:
          type: string
        purchaserContact:
          type: string
        pvCapacityKwp:
          type: number
        inverterModel:
          type: string
        inverterSerial:
          type: string
        moduleModel:
          type: string
        moduleSerial:
          type: string
        utilityAccountNumber:
          type: string
        meterNumber:
          type: string
        nemApplicationDate:
          type: string
        nemApprovalDate:
          type: string
        commissioningDate:
          type: string
        handoverDate:
          type: string
        warrantyStart:
          type: string
        warrantyEnd:
          type: string
        unitStatus:
          type: string
          enum: [not_started, design_confirmed, equipment_allocated, installation_in_progress, installation_completed, testing_completed, utility_submission_pending, utility_submission_submitted, utility_approval_received, meter_installed, commissioned, handover_pending, handed_over, under_warranty, closed]
    HousingUnitUpdate:
      type: object
      properties:
        unitNumber:
          type: string
        block:
          type: string
        phase:
          type: string
        purchaserName:
          type: string
        purchaserContact:
          type: string
        pvCapacityKwp:
          type: number
        inverterModel:
          type: string
        inverterSerial:
          type: string
        moduleModel:
          type: string
        moduleSerial:
          type: string
        utilityAccountNumber:
          type: string
        meterNumber:
          type: string
        nemApplicationDate:
          type: string
        nemApprovalDate:
          type: string
        commissioningDate:
          type: string
        handoverDate:
          type: string
        warrantyStart:
          type: string
        warrantyEnd:
          type: string
        unitStatus:
          type: string
          enum: [not_started, design_confirmed, equipment_allocated, installation_in_progress, installation_completed, testing_completed, utility_submission_pending, utility_submission_submitted, utility_approval_received, meter_installed, commissioned, handover_pending, handed_over, under_warranty, closed]